+++ src/core/algo.rs
@@ -506,7 +506,7 @@ fn install_dst(source_ip: &Ip, root: &PathBuf, mapping: &HashMap<LangIdentifier,
     // determine the cache slot name
     let cache_path = {
         let cache_slot = CacheSlot::new(
-            source_ip.get_man().get_ip().get_name(),
+            source_ip.get_uuid(),
             source_ip.get_man().get_ip().get_version(),
             &sum,
         );