+++ CHANGELOG.md
@@ -20,11 +20,14 @@ Add `- unreleased` along the next future version to prevent CI/CD from triggerin
 ### Features
 - supports finding ip using their name and their uuid in the catalog
 - adds uuid field to manifest
+- adds `ORBIT_IP_UUID` environment variable
 - copies the ip's lockfile to the channel now too (not just the manifest)
 - adds `--offline` flag to skip checking coherency with remote source when installing a local ip
 
 ### Changes
+- refactors how ips are stored in the archive and cache (uses uuid instead of name)
 - writes the full ip spec to the lock file (includes uuid)
+- extends checksum display length from 10 characters to 16 characters
 - requires the uuid field for ip manifests
 - reformats how uuid appears (uses base36 encoding instead of hyphenated format)
 - creates lockfile during `orbit new` and `orbit init` commands